Add NeutralColorEditor, separate from edit page
