+++ src/components/layout/BaseColors/index.ts
@@ -1,4 +1,5 @@
 export * from './ColorListItem';
 export * from './DebouncedColorPicker';
+export * from './NeutralColorEditor';
 export * from './PalettePreview';
 export * from './PrimaryColorEditor';
